Rewrite to generate separate locals and remotes lists
Also add code to correctly omit commands if there's nothing to do.

+++ stales.d
@@ -5,6 +5,7 @@ import std.datetime;
 import std.exception : enforce;
 import std.getopt;
 import std.process;
+import std.range : chain;
 import std.regex;
 import std.stdio;
 import std.string : strip, stripLeft, indexOf;
@@ -21,19 +22,20 @@ enum BranchType {
     LocalAndRemote = (Local | Remote)
 }
 
-struct BranchInfo {
-    string fullName;
-    BranchType type;
+struct StaleLists
+{
+    string[] locals;
+    string[] remotes;
 }
 
 /// Returns an array of branches that don't match the "keeper" regexes
-BranchInfo[] findRemoteBranches(string[] keepers, BranchType type)
+StaleLists getStaleBranches(string[] keepers, BranchType type)
 {
     // Compile our regular expressions first
     // (to spit out any regex errors before we light up "git branch")
     Regex!char[] regexes = keepers.map!(k => regex(k)).array;
 
-    BranchInfo[] ret;
+    StaleLists ret;
 
     if (type & BranchType.Local) {
         auto localBranchFinder = pipeProcess(["git", "branch"], Redirect.stdout);
@@ -51,12 +53,10 @@ BranchInfo[] findRemoteBranches(string[] keepers, BranchType type)
             .map!(b => stripLeft(b))
             // Filter out ones we want to keep
             .filter!(b => !regexes.any!(r => b.matchFirst(r)))
-            // Allocate a new string for each line (byLine reuses a buffer)
-            .map!(m => m.idup)
-            // Tag on the type
-            .map!(b => BranchInfo(b, BranchType.Local));
+            // Take a copy since byLine reuses the same buffer for each iteration.
+            .map!(b => b.idup);
 
-        ret = localBranches.array;
+        ret.locals = localBranches.array;
     }
 
     if (type & BranchType.Remote) {
@@ -74,29 +74,24 @@ BranchInfo[] findRemoteBranches(string[] keepers, BranchType type)
             .filter!(b => !b.canFind("->"))
             // The rest mostly the same as above
             .map!(b => stripLeft(b))
-            // We idup before we filter because extractBranchName expects
-            // an immutable string.
-            .map!(m => m.idup)
             .filter!(b => !regexes.any!(r => extractBranchName(b).matchFirst(r)))
-            .map!(b => BranchInfo(b, BranchType.Remote));
+            .map!(b => b.idup);
 
-        auto app = appender(ret);
-        app.put(remoteBranches);
-        ret = app.data;
+        ret.remotes = remoteBranches.array;
     }
 
     return ret;
 }
 
 /// Pulls the branch name off a "remote/branch" string
-string extractBranchName(string fullBranchName) pure
+auto extractBranchName(Range)(Range fullBranchName) pure
 {
     immutable firstSlash = fullBranchName.indexOf('/');
     return fullBranchName[firstSlash + 1 .. $];
 }
 
 /// Pulls the remote name off a "remote/branch" string
-string extractRemote(string fullBranchName) pure
+auto extractRemote(Range)(Range fullBranchName) pure
 {
     immutable firstSlash = fullBranchName.indexOf('/');
     return fullBranchName[0 .. firstSlash];
@@ -194,44 +189,43 @@ int main(string[] args)
     if (execute(["git", "rev-parse", mainBranch]).status != 0)
         writeAndFail("Git couldn't find the branch, \"", mainBranch, '"');
 
-    auto staleAppender = appender!(BranchInfo[])();
+    auto stales = getStaleBranches(keepers, bt);
 
-    // Get a list of branches we might be interested in,
-    // then build up a list of stale ones.
-    foreach (branch; findRemoteBranches(keepers, bt)) {
-        immutable currentName = branch.fullName;
-
+    bool filterStales(string branch) {
         // Ignore unmerged branches
-        immutable counts = aheadBehindCounts(currentName, mainBranch);
+        immutable counts = aheadBehindCounts(branch, mainBranch);
         if (counts.ahead > 0) {
             if (verbosity > 1) {
-                stderr.writeln(currentName, " has ", counts.ahead,
+                stderr.writeln(branch, " has ", counts.ahead,
                                " unmerged commits. Skipping.");
             }
-            continue;
+            return false;
         }
 
         // Ignore younger branches
         immutable ageInDays =
-                (branchTime(mainBranch) - branchTime(currentName)).total!"days";
+                (branchTime(mainBranch) - branchTime(branch)).total!"days";
         if (ageInDays < ageCutoff) {
             if (verbosity > 1) {
-                stderr.writeln(currentName, " is ", ageInDays,
+                stderr.writeln(branch, " is ", ageInDays,
                                " days old. Skipping.");
             }
-            continue;
+            return false;
         }
 
         // We have a keeper.
         if (verbosity > 0) {
-            stderr.writeln(currentName, " is ", ageInDays, " days old and ",
+            stderr.writeln(branch, " is ", ageInDays, " days old and ",
                            counts.behind, " commits behind ", mainBranch, '.');
         }
-        staleAppender.put(branch);
+        return true;
     }
-    BranchInfo[] stales = staleAppender.data();
 
-    if (stales.empty) {
+
+    stales.locals = stales.locals.filter!(filterStales).array;
+    stales.remotes = stales.remotes.filter!(filterStales).array;
+
+    if (stales.locals.empty && stales.remotes.empty) {
         writeln("No stale branches! (nothing to do)");
         return 0;
     }
@@ -239,61 +233,54 @@ int main(string[] args)
     // Build up commands to delete the stale branches if the user wants.
     if (dryRun || pushDeletes) {
 
-        // Local and remote branches shouldn't be randomly interspersed,
-        // and this makes our job easier below.
-        assert(isPartitioned!(s => s.type == BranchType.Local)(stales));
-
-        // First, the locals.
-        string[] localDeleteCommand = ["git", "branch", "-d"];
-        while (stales[0].type == BranchType.Local) {
-            localDeleteCommand ~= stales[0].fullName;
-            stales = stales[1 .. $];
-        }
-        if (pushDeletes) {
-            write("deleting local branches...");
-            stdout.flush();
-            auto deleteResult = execute(localDeleteCommand);
-            enforce(deleteResult.status == 0, "git branch -d failed");
-            writeln(" done");
-        }
-        else {
-            writeln(joiner(localDeleteCommand, " "));
-        }
-
-        // We should be done with local branches now
-        assert(all!(s => s.type == BranchType.Remote)(stales));
-
-        // Group branches by remote
-        // The sorting is probably unneeded - doesn't Git organize branches
-        // by remote?
-        auto remoteGroups = stales
-            .sort!((a, b) => extractRemote(a.fullName) < extractRemote(b.fullName))
-            .map!(s => s.fullName)
-            .chunkBy!(a => extractRemote(a));
-
-        foreach (group; remoteGroups) {
-            // group is a tuple. Member 0 is the remote name.
-            string[] remoteDeleteCommand = ["git", "push", "--delete", group[0]];
-            // Member 1 is the list of branches for that remote.
-            remoteDeleteCommand ~= group[1].map!(b => extractBranchName(b)).array;
+        if (!stales.locals.empty) {
+            // First, the locals.
+            string[] localDeleteCommand = ["git", "branch", "-d"] ~ stales.locals;
 
             if (pushDeletes) {
-                write("deleting remote branches...");
+                write("deleting local branches...");
                 stdout.flush();
-                auto deleteResult = execute(remoteDeleteCommand);
-                enforce(deleteResult.status == 0, "git push --delete failed");
+                auto deleteResult = execute(localDeleteCommand);
+                enforce(deleteResult.status == 0, "git branch -d failed");
                 writeln(" done");
             }
             else {
-                writeln(joiner(remoteDeleteCommand, " "));
+                writeln(joiner(localDeleteCommand, " "));
+            }
+        }
+
+        if (!stales.remotes.empty) {
+            // Group branches by remote
+            // The sorting is probably unneeded - doesn't Git organize branches
+            // by remote?
+            auto remoteGroups = stales.remotes
+                .sort
+                .chunkBy!(a => extractRemote(a));
+
+            foreach (group; remoteGroups) {
+                // group is a tuple. Member 0 is the remote name.
+                string[] remoteDeleteCommand = ["git", "push", "--delete", group[0]];
+                // Member 1 is the list of branches for that remote.
+                remoteDeleteCommand ~= group[1].map!(b => extractBranchName(b)).array;
+
+                if (pushDeletes) {
+                    write("deleting remote branches...");
+                    stdout.flush();
+                    auto deleteResult = execute(remoteDeleteCommand);
+                    enforce(deleteResult.status == 0, "git push --delete failed");
+                    writeln(" done");
+                }
+                else {
+                    writeln(joiner(remoteDeleteCommand, " "));
+                }
             }
         }
     }
     // Otherwise just print them out.
     else {
         writeln("Stale branches:");
-        foreach (stale; stales)
-            writeln(stale.fullName);
+        foreach (stale; chain(stales.locals, stales.remotes))
+            writeln(stale);
     }
     return 0;
 }
